+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `compote version` command to print the version of the component manager
 - Add `IDF_COMPONENT_OVERWRITE_MANAGED_COMPONENTS` environment variable to allow overwrite files in the managed_component directory, even if they have been modified by the user
 - Add documentation project for the component manager and Espressif component registry
+- Treat local source dependency priority higher
 
 ### Fixed